# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CryptoCampus API
  version: 1.0.0
  description: |
    团队 01 的 HTTP 接口唯一机器可读契约。所有业务接口以 /api/v1 为前缀。
    修改字段、类型、错误码或权限前，必须先评审本文件。
...
servers:
  - url: /api/v1
    description: 当前部署
...
paths:
...
  /me:
    get:
      tags: [Me]
      operationId: getMe
      summary: 获取当前用户
      x-required-roles: [student, admin, teacher]
      x-jwt: true
      x-sensitive-fields: [email]
      responses:
        "200":
          description: 当前用户
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /me/keyring:
    get:
      tags: [Me]
      operationId: getKeyring
      summary: 获取密钥指纹和证书状态，不返回私钥
      x-required-roles: [student, admin, teacher]
      x-jwt: true
      x-sensitive-fields: [private_key, kek]
      responses:
        "200":
          description: 密钥环摘要
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyringSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/certificate/verify:
    get:
      tags: [Me]
      operationId: verifyOwnCertificate
      summary: 使用平台 CA 和最新 CRL 验证当前用户证书链
      x-required-roles: [student, admin, teacher]
      x-jwt: true
      responses:
        "200":
          description: 当前用户证书链验证结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CertificateVerificationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /me/keyring/rotate:
    post:
      tags: [Me]
      operationId: rotateKeyring
      summary: 二次口令确认后轮换密钥并吊销旧证书
      x-required-roles: [student, admin, teacher]
      x-jwt: true
      x-sensitive-fields: [password, private_key, kek]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [password, acknowledge_inflight_loss]
              properties:
                password: { type: string, writeOnly: true }
                acknowledge_inflight_loss: { const: true }
      responses:
        "200":
          description: 新密钥环摘要
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KeyringSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
  /me/sessions:
    get:
      tags: [Me]
      operationId: listSessions
      summary: 获取登录设备列表
      x-required-roles: [student, admin, teacher]
      x-jwt: true
      x-sensitive-fields: [ip]
      responses:
        "200":
          description: 会话列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/DeviceSession" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me/sessions/{session_id}:
    delete:
      tags: [Me]
      operationId: revokeSession
      summary: 远程登出指定设备
      x-required-roles: [student, admin, teacher]
      x-jwt: true
      x-sensitive-fields: [refresh_token]
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "204": { description: 已吊销 }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /me/quotas:
    get:
      tags: [Me]
      operationId: getQuotas
      summary: 获取当前 UTC 日的可展示服务额度
      description: 仅返回按日计数的树洞凭证、互动凭证和密信创建额度；投票凭证按 vote_id 计数，不适合汇总为单一当日额度。
      x-required-roles: [student, admin, teacher]
      x-jwt: true
      x-sensitive-fields: []
      responses:
        "200":
          description: 额度
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QuotaPage" }
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /me/keyring/export:
    post:
      tags: [Me]
      operationId: exportKeyringBackup
      summary: 导出口令加密的 PEM 密钥备份
      x-required-roles: [student, admin, teacher]
      x-jwt: true
      x-sensitive-fields: [password, private_key, backup_file]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [password]
              properties:
                password: { type: string, minLength: 1, maxLength: 128, writeOnly: true }
      responses:
        "200":
          description: 口令加密的 PEM 文件
          content:
            application/x-pem-file:
              schema: { type: string, format: binary }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
  /me/keyring/import:
    post:
      tags: [Me]
      operationId: importKeyringBackup
      summary: 导入口令加密的 PEM 备份并恢复密钥环
      x-required-roles: [student, admin, teacher]
      x-jwt: true
      x-sensitive-fields: [password, backup_file, private_key, kek]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [backup, password]
              properties:
                backup: { type: string, format: binary }
                password: { type: string, minLength: 1, maxLength: 128, writeOnly: true }
      responses:
        "200":
          description: 恢复后的密钥环摘要
          content:
            application/json:
              schema: { $ref: "#/components/schemas/KeyringSummary" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
  /me/notifications:
    get:
      tags: [Me]
      operationId: listNotifications
      summary: 获取本人站内通知，包括密信提取通知
      x-required-roles: [student, admin, teacher]
      x-jwt: true
      x-sensitive-fields: [drop_plaintext]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: 通知分页列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /me/account:
    delete:
      tags: [Me]
      operationId: deleteAccount
      summary: 注销本人账号并执行课程规定的数据删除与保留策略
      x-required-roles: [student, admin, teacher]
      x-jwt: true
      x-sensitive-fields: [password, private_key, refresh_token]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [password, confirm]
              properties:
                password: { type: string, minLength: 1, maxLength: 128, writeOnly: true }
                confirm: { const: DELETE_MY_ACCOUNT }
      responses:
        "202":
          description: 注销流程已接受
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Accepted" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: cc_refresh
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    SessionId:
      name: session_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: 未认证、凭据错误或会话失效
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: 资源不存在或为防枚举隐藏存在性
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: 状态冲突、重复消费或幂等冲突
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    PayloadTooLarge:
      description: 文件超过当前操作限制；密信文件的课程上限为 100 MiB
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServiceUnavailable:
      description: 密码引擎或 Provider 暂不可用
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    BadRequest:
      description: 请求语义、密码材料或状态不合法
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InternalError:
      description: 后端或密码引擎内部错误；消息不得包含密钥、明文或原始堆栈
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Email:
      type: string
      format: email
      maxLength: 254
      description: 规范化为小写并去除首尾空白的校园邮箱
    Role:
      type: string
      enum: [student, admin, teacher, system]
    UserStatus:
      type: string
      enum: [active, frozen, pending_deletion]
    Accepted:
      type: object
      additionalProperties: false
      required: [accepted]
      properties:
        accepted: { const: true }
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id, details]
      properties:
        code: { type: string, pattern: "^[A-Z][A-Z0-9_]+$" }
        message: { type: string }
        request_id: { type: string }
        details: { type: object, additionalProperties: true }
...
    User:
      type: object
      additionalProperties: false
      required: [id, email, role, status, pqc_mode, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { $ref: "#/components/schemas/Email" }
        role: { $ref: "#/components/schemas/Role" }
        status: { $ref: "#/components/schemas/UserStatus" }
        pqc_mode: { type: boolean }
        created_at: { type: string, format: date-time }
...
    KeyringItem:
      type: object
      additionalProperties: false
      required: [kind, algorithm, fingerprint, status]
      properties:
        kind: { type: string, enum: [sm2_identity, certificate, ml_kem] }
        algorithm: { type: string }
        fingerprint: { type: string, description: SM3 指纹的脱敏展示 }
        status: { type: string, enum: [active, expired, revoked, unavailable] }
        expires_at: { type: [string, "null"], format: date-time }
    KeyringSummary:
      type: object
      additionalProperties: false
      required: [items, unlocked_until]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/KeyringItem" }
        unlocked_until: { type: [string, "null"], format: date-time }
    CertificateVerificationResponse:
      type: object
      additionalProperties: false
      required: [valid, state, serial, verified_at]
      properties:
        valid: { type: boolean }
        state: { type: string, enum: [active, expired, revoked, invalid] }
        serial: { type: [string, "null"] }
        verified_at: { type: string, format: date-time }
    DeviceSession:
      type: object
      additionalProperties: false
      required: [id, device, ip_masked, last_active_at, current]
      properties:
        id: { type: string, format: uuid }
        device: { type: string }
        ip_masked: { type: string }
        last_active_at: { type: string, format: date-time }
        current: { type: boolean }
    Quota:
      type: object
      additionalProperties: false
      required: [resource, used, limit]
      properties:
        resource: { type: string, enum: [hole_credential, interaction_credential, drop] }
        used: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1 }
    QuotaPage:
      type: object
      additionalProperties: false
      required: [items, resets_at]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Quota" }
        resets_at: { type: string, format: date-time }
...
    Notification:
      type: object
      additionalProperties: false
      required: [id, type, title, read, created_at]
      properties:
        id: { type: string, format: uuid }
        type: { type: string, enum: [drop_extracted, security, system] }
        title: { type: string, minLength: 1, maxLength: 200 }
        related_resource_id: { type: [string, "null"] }
        read: { type: boolean }
        created_at: { type: string, format: date-time }
    NotificationPage:
      type: object
      additionalProperties: false
      required: [items, page, page_size, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Notification" } }
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
...
security:
  - bearerAuth: []